+++ lib/compatible/ym_compatible_html.dart
@@ -1,0 +1,11 @@
+
+
+class ImageElement{
+
+  String src = "";
+
+}
+
+class DivElement{
+
+}