+++ lib/compatible/ym_compatible_io.dart
@@ -1,0 +1,7 @@
+
+class Platform{
+  static bool isIOS = true;
+  static bool isAndroid = true;
+}
+
+