+++ lib/compatible/ym_compatible_ui.dart
@@ -1,0 +1,5 @@
+
+
+class platformViewRegistry {
+  static registerViewFactory(String viewId, dynamic cb) {}
+}