+++ lib/compatible/ym_constants.dart
@@ -1,0 +1,7 @@
+
+enum  YmPlatform {
+  web,
+  android,
+  ios,
+  pc
+}