+++ lib/widgets/ListView/ym_list_view.dart
@@ -11,6 +11,7 @@ class YmListView extends StatefulWidget {
   final Size size;
   bool hasMoreData;
   late String loadMoreText;
+  final Color textColor;
 
   YmListView(
     this.items,
@@ -19,8 +20,10 @@ class YmListView extends StatefulWidget {
     this.size = const Size(500, 500),
     this.hasMoreData = true,
     this.loadMoreText = "没有更多了~",
+    this.textColor = const Color(0xFF606FFF),
     required this.onItemBuilder,
     required this.onLoadMore,
+
   }) : super(key: key);
 
   @override
@@ -96,7 +99,7 @@ class _YmListViewState extends State<YmListView> with SingleTickerProviderStateM
       return Container(
         width: widget.size.width,
         height: widget.size.height - 50,
-        child: YmEmpty("暂无数据", "assets/images/ic_no_data.png"),
+        child: YmEmpty("暂无数据", "assets/images/ic_no_data.png",textColor:widget.textColor),
       );
     }
   }